refactor: 리뷰 반영

+++ src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -2,21 +2,19 @@ package kitchenpos.product.dto;
 
 import kitchenpos.product.domain.Product;
 
-import java.math.BigDecimal;
-
 public class ProductRequest {
     private String name;
-    private BigDecimal price;
+    private Long price;
 
     protected ProductRequest() {
     }
 
-    public ProductRequest(String name, BigDecimal price) {
+    public ProductRequest(String name, Long price) {
         this.name = name;
         this.price = price;
     }
 
-    public static ProductRequest of(String name, BigDecimal price) {
+    public static ProductRequest of(String name, Long price) {
         return new ProductRequest(name, price);
     }
 
@@ -24,7 +22,7 @@ public class ProductRequest {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
